+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## 0.4.0
+
 - Add brace matcher for `[]`, `{}` and `()`
 - Fix completion of items with `/`